rename generic2 -> generic
