+++ polars/polars-lazy/polars-pipe/src/executors/sinks/groupby/generic/mod.rs
