+++ polars/polars-lazy/polars-pipe/src/executors/sinks/groupby/generic/thread_local.rs
